ca_spin_v2: showcase to change CP setting online
Example about how to change a task cache partition configuration while
the task is running.

The showcase is in the job()

NB: We cannot allow a program T1 to change the cache partition
configuration of another program T2 right now. This is because the
clflush can only run under the target task (i.e., T2) execution context.
Otherwise, the linear address to physical address translation in the HW
will use the wrong CR3 content.

Test
    Works on the cat-haswell machine.
    The task will change its cache partition configuration in each job.

How to test
    setsched PSN-EDF
    ./ca_spin_v2 -p 1 -s 4096 -l 1000 -q 1 -C 8 -P 0xff -D 2000 1000 2000 10